+++ src/compiler/binder/mod.rs
@@ -2,5 +2,9 @@ mod lines_0_500;
 pub use lines_0_500::*;
 mod lines_500_1000;
 pub use lines_500_1000::*;
-mod lines_1000_2000;
-pub use lines_1000_2000::*;
+mod lines_1000_1500;
+pub use lines_1000_1500::*;
+mod lines_1500_2000;
+pub use lines_1500_2000::*;
+mod lines_2000_3000;
+pub use lines_2000_3000::*;
